LPS-73791 match package com.liferay.item.selector.ItemSelectorReturnTypeResolver
